# Lab3/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Notes app",
    "description": "Simple app for manage notes"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [],
  "securityDefinitions": {},
  "paths": {
    "/api/v1/user": {
      "get": {
        "description": "Get all users",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of all users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Add user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a user",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Joe"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "user": {
                  "$ref": "#/definitions/User"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "User already exist",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "user already exist"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "username error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "username not supplied"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/v1/user/{userID}": {
      "get": {
        "description": "Get user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "user not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "user not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "description": "Edit user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "New name",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Joe"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "user": {
                  "$ref": "#/definitions/User"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "User already exist",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "user already exist"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "user not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "username error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "username not supplied"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/v1/note": {
      "get": {
        "description": "Get notes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "noteID",
            "description": "Note id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "path",
            "description": "Numper of page",
            "required": false,
            "type": "string"
          },
          {
            "name": "items_per_page",
            "in": "path",
            "description": "Items per page",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Note",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Note"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Add note",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a note",
            "schema": {
              "$ref": "#/definitions/Note_nid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "note": {
                  "$ref": "#/definitions/Note"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "note error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "noteID not supplied"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/link": {
      "post": {
        "description": "Add link",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Note id",
            "schema": {
              "$ref": "#/definitions/Link_nid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "link": {
                  "$ref": "#/definitions/Link"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "noteID error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "noteID not supplied"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "userID": {
          "type": "string",
          "example": "dbdb1e30-0078-11ee-ba2d-e5ada0d4a151"
        },
        "username": {
          "type": "string",
          "example": "Ivan"
        }
      }
    },
    "Note": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "dbdb1e31-0078-11ee-ba2d-e5ada0d4a151"
        },
        "userID": {
          "type": "string",
          "example": "dbdb1e32-0078-11ee-ba2d-e5ada0d4a151"
        },
        "title": {
          "type": "string",
          "example": "Title"
        },
        "note": {
          "type": "string",
          "example": "Content"
        }
      }
    },
    "Note_nid": {
      "type": "object",
      "properties": {
        "userID": {
          "type": "string",
          "example": "dbdb1e33-0078-11ee-ba2d-e5ada0d4a151"
        },
        "title": {
          "type": "string",
          "example": "Title"
        },
        "note": {
          "type": "string",
          "example": "Content"
        }
      }
    },
    "Link": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "dbdb1e34-0078-11ee-ba2d-e5ada0d4a151"
        },
        "noteID": {
          "type": "string",
          "example": "dbdb1e35-0078-11ee-ba2d-e5ada0d4a151"
        }
      }
    },
    "Link_nid": {
      "type": "object",
      "properties": {
        "noteID": {
          "type": "string",
          "example": "dbdb1e36-0078-11ee-ba2d-e5ada0d4a151"
        }
      }
    }
  }
}